[NCL-8063]: Remove EnumSetToStringConverter due to Deployment Exception
